Added SerializationManager and replace logic from ClanManager to it
