+++ src/main/java/io/github/lofrol/UselessClan/ClanConfigManager.java
@@ -1,0 +1,4 @@
+package io.github.lofrol.UselessClan;
+
+public class ClanConfigManager {
+}
