+++ src/main/java/io/github/lofrol/UselessClan/ClanManager.java
@@ -14,6 +14,7 @@ import java.io.IOException;
 import java.util.*;
 import java.util.logging.Level;
 
+import static io.github.lofrol.UselessClan.SerializationManager.*;
 import static org.bukkit.Bukkit.getServer;
 
 public final class ClanManager {
@@ -30,10 +31,6 @@ public final class ClanManager {
             "&5",       //9 lvl
             "&0",       //10 lvl
     };
-
-    private static final String ClanFolderName = "Clans";
-    private static final String DeletedClanFolder = "DeletedClans";
-    private static final String backupClanFolder = "backups";
 
     private final UselessClan OwnerPlugin;
 
@@ -96,7 +93,7 @@ public final class ClanManager {
      */
     public void LoadClans() {
 
-        File tempDir = checkClanFolderOrCreate(ClanFolderName);
+        File tempDir = UselessClan.getSerilManager().checkClanFolderOrCreate(ClanFolderName);
 
         for (File tempClanFile : Objects.requireNonNull(tempDir.listFiles())) {
             FileConfiguration ClanConfig = new YamlConfiguration();
@@ -118,7 +115,7 @@ public final class ClanManager {
 
     public void SaveClans() {
 
-        File tempDir = checkClanFolderOrCreate(ClanFolderName);
+        File tempDir = UselessClan.getSerilManager().checkClanFolderOrCreate(ClanFolderName);
 
         OwnerPlugin.getLogger().log(Level.INFO, "Starting save clans to plugin folder...");
         for (Clan TempClan : ServerClans.values()) {
@@ -155,8 +152,8 @@ public final class ClanManager {
     }
 
     private void RemoveClanToDeletedClanFolder(Clan clanToDelete) {
-        File tempDeleteDir = checkClanFolderOrCreate(DeletedClanFolder);
-        File tempClanDir = checkClanFolderOrCreate(ClanFolderName);
+        File tempDeleteDir = UselessClan.getSerilManager().checkClanFolderOrCreate(DeletedClanFolder);
+        File tempClanDir = UselessClan.getSerilManager().checkClanFolderOrCreate(ClanFolderName);
 
         FileConfiguration ClanConfig = clanToDelete.SaveClanToConfig();
         boolean isDeleted = false;
@@ -190,27 +187,6 @@ public final class ClanManager {
         OwnerPlugin.getLogger().log(Level.FINE, String.format("%s was deleted successfully", clanToDelete.getPrefixClan()));
     }
 
-
-    // Return Clans folder
-    private File checkClanFolderOrCreate(String FolderName) {
-        if (!checkFolderOrCreate(OwnerPlugin.getDataFolder())) {
-            OwnerPlugin.getLogger().log(Level.SEVERE, "Cant create plugin folder!");
-        }
-
-        File tempDir = new File(OwnerPlugin.getDataFolder(), FolderName);
-
-        if (checkFolderOrCreate(tempDir)) {
-            OwnerPlugin.getLogger().log(Level.INFO, String.format("Checked %s folder for exist", FolderName));
-        }
-        else {
-            OwnerPlugin.getLogger().log(Level.SEVERE, String.format("Cant create %s folder!", FolderName));
-        }
-        return tempDir;
-    }
-
-    private boolean checkFolderOrCreate(File Folder) {
-        return (Folder.exists() || Folder.mkdir());
-    }
 
 
     /*
@@ -267,8 +243,7 @@ public final class ClanManager {
                 ChatSender.MessageTo(player,"UselessClan",
                         String.format("Your clan have %d requests for join! ./clan requests", tempClan.getRequestCount()));
             }
-        }
-                , 200);
+        }, 200);
 
         OwnerPlugin.getLogger().log(Level.INFO, String.format( "Clan member %s Join to server, his clan is %s", player.getName(), tempClan.getNameClan()));
     }
@@ -284,7 +259,7 @@ public final class ClanManager {
 
     public void createClansBackups() {
         OwnerPlugin.getLogger().log(Level.INFO, "Starting backup of server clans...");
-        File tempDir = checkClanFolderOrCreate(backupClanFolder);
+        File tempDir = UselessClan.getSerilManager().checkClanFolderOrCreate(backupClanFolder);
 
         Calendar tempDate = Calendar.getInstance();
 
@@ -292,12 +267,8 @@ public final class ClanManager {
                 tempDate.get(Calendar.DAY_OF_MONTH), tempDate.get(Calendar.MONTH), tempDate.get(Calendar.YEAR),
                 tempDate.get(Calendar.HOUR), tempDate.get(Calendar.MINUTE)));
 
-        if (!todayBackupFile.exists()) {
-            if (todayBackupFile.mkdir()) {
-                OwnerPlugin.getLogger().log(Level.INFO, "Creating folder for backups...");
-            }
-            return;
-        }
+        UselessClan.getSerilManager().checkFolderOrCreate(todayBackupFile);
+
         for (Clan tempClan : ServerClans.values()) {
             File tempClanFile = new File(todayBackupFile, String.format("%s.yml", tempClan.getPrefixClan()));
             FileConfiguration tempConfig = tempClan.SaveClanToConfig();
