--- src/main/java/io/github/lofrol/UselessClan/Extensions/LocalizationManager.java
@@ -1,4 +1,0 @@
-package io.github.lofrol.UselessClan.Extensions;
-
-public class LocalizationManager {
-}
